shortener/serializers.py
# serializers.py in shortener app
from rest_framework import serializers

from .models import ShortURL

class ShortURLSerializer(serializers.ModelSerializer):
    class Meta:
        model = ShortURL
        fields = ['original_url']
