requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.1.2
djangorestframework==3.15.2
h11==0.14.0
sqlparse==0.5.1
uvicorn==0.31.1
gunicorn
psycopg2-binary==2.9.9
